Core CaSMAI script
This is the Core CaSMAI script. Due note, this doesn't include all of the required services to run the actual script. To stay within GIThub's terms of service, those will have to be procured separately. But they can be found in the model imports and the list of required llama models. This script largely employs Ollama.